requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
boto3==1.22.9
botocore==1.25.9
bottleneck==1.3.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dill==0.3.4
docker-pycreds==0.4.0
elasticdeform==0.4.9
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gitdb==4.0.9
gitpython==3.1.27
google-auth-oauthlib==0.4.6
google-auth==2.6.0
grpcio==1.44.0
hub==2.4.1
humbug==0.2.7
idna==3.3
imageio==2.16.0
importlib-metadata==4.11.3
itk-core==5.2.1.post1
itk-elastix==0.13.0
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
itk==5.2.1.post1
jmespath==1.0.0
joblib==1.1.0
kaggle==1.5.12
kiwisolver==1.3.2
kornia==0.6.4
llvmlite==0.38.0
lpips==0.1.4
markdown==3.3.6
matplotlib==3.5.1
medpy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==0.8.1
multidict==6.0.2
multiprocess==0.70.12.2
munkres==1.1.4
networkx==2.6.3
nibabel==3.2.2
numba==0.55.1
numcodecs==0.9.1
numexpr==2.8.1
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pathos==0.2.8
pathtools==0.1.2
pillow==9.0.1
pip==21.2.4
plotly==5.6.0
pox==0.3.0
ppft==1.6.6.4
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydeprecate==0.3.1
pynndescent==0.5.6
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.2
pytorch-fid==0.2.1
pytorch-ignite==0.4.9
pytorch-lightning==1.4.9
pytorch-msssim==0.2.1
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
requests-oauthlib==1.3.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sentry-sdk==1.5.6
setproctitle==1.2.2
setuptools==59.5.0
shortuuid==1.0.8
simpleitk==2.1.1
sip==4.19.13
six==1.16.0
smmap==5.0.0
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.2.9
torch-dct==0.1.5
torchgeometry==0.1.2
torchmetrics==0.7.2
torchsummary==1.5.1
tornado==6.1
tqdm==4.62.3
typing-extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.8
wandb==0.12.11
werkzeug==2.0.3
wheel==0.37.1
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0